Update to follow proposed PayoutFunction spec
+ Add HyperbolaPayoutCurvePiece
