+++ packages/messaging/lib/messages/PayoutFunction.ts
@@ -1,9 +1,11 @@
 import { BufferReader, BufferWriter } from '@node-lightning/bufio';
 import { MessageType } from '../MessageType';
+import { getTlv } from '../serialize/getTlv';
 import { IDlcMessage } from './DlcMessage';
+import { PayoutCurvePiece } from './PayoutCurvePiece';
 
 export abstract class PayoutFunction {
-  public static deserialize(buf: Buffer): PayoutFunctionV0 | PayoutFunctionV1 {
+  public static deserialize(buf: Buffer): PayoutFunctionV0 {
     const reader = new BufferReader(buf);
 
     const type = Number(reader.readBigSize());
@@ -11,12 +13,8 @@ export abstract class PayoutFunction {
     switch (type) {
       case MessageType.PayoutFunctionV0:
         return PayoutFunctionV0.deserialize(buf);
-      case MessageType.PayoutFunctionV1:
-        return PayoutFunctionV1.deserialize(buf);
       default:
-        throw new Error(
-          `Payout function TLV type must be PayoutFunctionV0 or PayoutFunctionV1`,
-        );
+        throw new Error(`Payout function TLV type must be PayoutFunctionV0`);
     }
   }
 
@@ -43,18 +41,21 @@ export class PayoutFunctionV0 extends PayoutFunction implements IDlcMessage {
 
     reader.readBigSize(); // read type
     instance.length = reader.readBigSize(); // need to fix this
-    reader.readUInt16BE(); // num_pts
+    reader.readUInt16BE(); // num_pieces
+    instance.endpoint0 = reader.readBigSize();
+    instance.endpointPayout0 = reader.readBigSize();
+    instance.extraPrecision0 = reader.readUInt16BE();
 
     while (!reader.eof) {
-      const isEndpoint = reader.readUInt8() === 1;
-      const eventOutcome = reader.readBigSize();
-      const outcomePayout = reader.readBigSize();
+      const payoutCurvePiece = PayoutCurvePiece.deserialize(getTlv(reader));
+      const endpoint = reader.readBigSize();
+      const endpointPayout = reader.readBigSize();
       const extraPrecision = reader.readUInt16BE();
 
-      instance.points.push({
-        isEndpoint,
-        eventOutcome,
-        outcomePayout,
+      instance.pieces.push({
+        payoutCurvePiece,
+        endpoint,
+        endpointPayout,
         extraPrecision,
       });
     }
@@ -69,7 +70,11 @@ export class PayoutFunctionV0 extends PayoutFunction implements IDlcMessage {
 
   public length: bigint;
 
-  public points: IPoint[];
+  public endpoint0: bigint;
+  public endpointPayout0: bigint;
+  public extraPrecision0: number;
+
+  public pieces: IPayoutCurvePieces[];
 
   /**
    * Serializes the payout_function_v0 message into a Buffer
@@ -78,131 +83,19 @@ export class PayoutFunctionV0 extends PayoutFunction implements IDlcMessage {
     const writer = new BufferWriter();
     writer.writeBigSize(this.type);
     writer.writeBigSize(this.length);
-    writer.writeUInt16BE(this.points.length);
-
-    for (const point of this.points) {
-      writer.writeUInt8(point.isEndpoint ? 1 : 0);
-      writer.writeBigSize(point.eventOutcome);
-      writer.writeBigSize(point.outcomePayout);
-      writer.writeUInt16BE(point.extraPrecision);
-    }
-
-    return writer.toBuffer();
-  }
-}
-
-/**
- * PayoutFunction V1
- */
-export class PayoutFunctionV1 extends PayoutFunction implements IDlcMessage {
-  public static type = MessageType.PayoutFunctionV1;
-
-  /**
-   * Deserializes an payout_function_v1 message
-   * @param buf
-   */
-  public static deserialize(buf: Buffer): PayoutFunctionV1 {
-    const instance = new PayoutFunctionV1();
-    const reader = new BufferReader(buf);
-
-    reader.readBigSize(); // read type
-    instance.length = reader.readBigSize(); // need to fix this
-    instance.usePositivePiece = reader.readUInt8() === 1;
-    instance.translateOutcomeSign = reader.readUInt8() === 1;
-    instance.translateOutcome = reader.readBigSize();
-    instance.translateOutcomeExtraPrecision = reader.readUInt16BE();
-    instance.translatePayoutSign = reader.readUInt8() === 1;
-    instance.translatePayout = reader.readBigSize();
-    instance.translatePayoutExtraPrecision = reader.readUInt16BE();
-    instance.aSign = reader.readUInt8() === 1;
-    instance.a = reader.readBigSize();
-    instance.aExtraPrecision = reader.readUInt16BE();
-    instance.bSign = reader.readUInt8() === 1;
-    instance.b = reader.readBigSize();
-    instance.bExtraPrecision = reader.readUInt16BE();
-    instance.cSign = reader.readUInt8() === 1;
-    instance.c = reader.readBigSize();
-    instance.cExtraPrecision = reader.readUInt16BE();
-    instance.dSign = reader.readUInt8() === 1;
-    instance.d = reader.readBigSize();
-    instance.dExtraPrecision = reader.readUInt16BE();
-
-    return instance;
-  }
-
-  /**
-   * The type for payout_function_v1 message. payout_function_v1 = 42792
-   */
-  public type = PayoutFunctionV1.type;
-
-  public length: bigint;
-
-  public usePositivePiece: boolean;
-
-  public translateOutcomeSign: boolean;
-
-  public translateOutcome: bigint;
-
-  public translateOutcomeExtraPrecision: number;
-
-  public translatePayoutSign: boolean;
-
-  public translatePayout: bigint;
-
-  public translatePayoutExtraPrecision: number;
-
-  public aSign: boolean;
-
-  public a: bigint;
-
-  public aExtraPrecision: number;
-
-  public bSign: boolean;
-
-  public b: bigint;
-
-  public bExtraPrecision: number;
-
-  public cSign: boolean;
-
-  public c: bigint;
-
-  public cExtraPrecision: number;
-
-  public dSign: boolean;
-
-  public d: bigint;
-
-  public dExtraPrecision: number;
-
-  /**
-   * Serializes the enum_event_descriptor_v0 message into a Buffer
-   */
-  public serialize(): Buffer {
-    const writer = new BufferWriter();
-    writer.writeBigSize(this.type);
 
     const dataWriter = new BufferWriter();
+    dataWriter.writeUInt16BE(this.pieces.length);
+    dataWriter.writeBigSize(this.endpoint0);
+    dataWriter.writeBigSize(this.endpointPayout0);
+    dataWriter.writeUInt16BE(this.extraPrecision0);
 
-    dataWriter.writeUInt8(this.usePositivePiece ? 1 : 0);
-    dataWriter.writeUInt8(this.translateOutcomeSign ? 1 : 0);
-    dataWriter.writeBigSize(this.translateOutcome);
-    dataWriter.writeUInt16BE(this.translateOutcomeExtraPrecision);
-    dataWriter.writeUInt8(this.translatePayoutSign ? 1 : 0);
-    dataWriter.writeBigSize(this.translatePayout);
-    dataWriter.writeUInt16BE(this.translatePayoutExtraPrecision);
-    dataWriter.writeUInt8(this.aSign ? 1 : 0);
-    dataWriter.writeBigSize(this.a);
-    dataWriter.writeUInt16BE(this.aExtraPrecision);
-    dataWriter.writeUInt8(this.bSign ? 1 : 0);
-    dataWriter.writeBigSize(this.b);
-    dataWriter.writeUInt16BE(this.bExtraPrecision);
-    dataWriter.writeUInt8(this.cSign ? 1 : 0);
-    dataWriter.writeBigSize(this.c);
-    dataWriter.writeUInt16BE(this.cExtraPrecision);
-    dataWriter.writeUInt8(this.dSign ? 1 : 0);
-    dataWriter.writeBigSize(this.d);
-    dataWriter.writeUInt16BE(this.dExtraPrecision);
+    for (const piece of this.pieces) {
+      dataWriter.writeBytes(piece.payoutCurvePiece.serialize());
+      dataWriter.writeBigSize(piece.endpoint);
+      dataWriter.writeBigSize(piece.endpointPayout);
+      dataWriter.writeUInt16BE(piece.extraPrecision);
+    }
 
     writer.writeBigSize(dataWriter.size);
     writer.writeBytes(dataWriter.toBuffer());
@@ -211,9 +104,9 @@ export class PayoutFunctionV1 extends PayoutFunction implements IDlcMessage {
   }
 }
 
-interface IPoint {
-  isEndpoint: boolean;
-  eventOutcome: bigint;
-  outcomePayout: bigint;
+interface IPayoutCurvePieces {
+  payoutCurvePiece: PayoutCurvePiece;
+  endpoint: bigint;
+  endpointPayout: bigint;
   extraPrecision: number;
 }
